nga-curator/requirements.txt
streamlit
pandas
google-cloud-aiplatform
google-cloud-bigquery
pandas-gbq
pytest
pytest-html
